Refactor model metadata table into ModelMetadata component
